Restructure api package
