+++ go/resource/deepcopy_gen.sh
