+++ go/resource/training_job_test.go
